+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# cdata 1.1.0 2019/04/21
+
+ * General transform specification.
+
 # cdata 1.0.9 2019/04/20
 
  * "layout" commands.